Fix margin trading logic for buy/sell orders
- Update ProcessSubmit in internal/api/store_engine.go to correctly calculate and lock initial margin (quote currency) based on the leverage instead of the full notional value or base asset quantity for both margin BUY and SELL orders (leverage > 1).
- Fix Python test script argument parsing to send price as an integer to prevent JSON unmarshaling errors in the Go backend.
- Remove server_bin and add to .gitignore alongside Python venv files.

+++ scripts/python_client/run_orders.py
@@ -19,7 +19,7 @@ def main():
     parser.add_argument("--side", choices=["BUY", "SELL"], default="BUY", help="Order Side")
     parser.add_argument("--type", choices=["MARKET", "LIMIT", "STOP", "STOP_LIMIT"], default="LIMIT", help="Order Type")
     parser.add_argument("--quantity", type=int, default=10, help="Order Quantity")
-    parser.add_argument("--price", type=float, help="Order Price")
+    parser.add_argument("--price", type=int, help="Order Price")
     parser.add_argument("--leverage", type=int, default=1, help="Leverage")
 
     args = parser.parse_args()
